workspace/paper-repos/paper_repo/requirements.txt
numpy>=1.24.0
scipy>=1.11.0
torch>=2.1.0
pandas>=2.0.0
matplotlib>=3.7.0
pyyaml>=6.0
tqdm>=4.65.0
signatory==1.2.7.1.9.0
